backend/pyproject.toml
[project]
name = "data-governance-backend"
version = "0.1.0"
description = "Data Governance Chat Application - FastAPI + LangGraph backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "langgraph>=0.4",
    "langchain-openai>=0.3",
    "langchain-core>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-chroma>=0.2",
    "langchain-community>=0.3",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "mcp>=1.9",
    "pydantic-settings>=2.7",
    "sse-starlette>=2.2",
    "httpx>=0.28",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
